Handle double-open of cameras better
This is a case that should never, ever happen, but potentially can
happen in certain error conditions where some events that lead to the
closing of the camera are not processed before something tries to open
the camera again.
In any case, these errors should be found and fixed, this change makes
finding the problems easier.