Transparently reorder module initialization to be more robust
With this change, frame sources will always be initialized before sinks
and scripts will always be run last.
This should add a bit of determinism for modules and allow them to make
assumptions about their sources in their preparation step.
(In particular, a VideoWriter injected into a frame source will now no
longer be reset if the writer is initialized before the frame source is
ready)