Create ModuleInfo interface to make the actual modules opaque in future
This will - in the long run - make module implementation easier by
making the module implementations and declarations more opaque from the
manager.
Additionally, the module machanics themselves are simplified, e.g. it is
now permitted to initialize functionally relevant stuff in a module
constructor. This will make a lot of "Are we initialized already?"
checks disappear and hopefully lead to less errors.