Synchronize thread processing start with wait conditions
The wait conditio is optional, as a module may also just directly set
its state to ready and thereby clear the requirement for the experiment
to run. Using the mutex / wait condition is more efficient than the
previous commonplace busy-looping, which - depending on the module
arrangement and scheduler - created load spikes during initialization.