Make engine aware of OOP modules and treat them specially
The engine now waits for OOP modules to become ready properly before
continuing, and also bunches them all together into one thread, as
communication with OOP modules is a pretty light computation task and we
don't need to run extra threads just for that and unnecessarily keep the
CPU busy. In case there is a performance limit later, we will always be
able to adjust this easily.