Initial work on EDL port to TOML and data storage improvements
To bring EDL a bit closer to the exDir specification (which I just
recently learned about), it is ported from JSON to TOML and also split
from the module API to be a lot more generic and potentially useful for
more than just Syntalos.
The current reimplementation will eventually supersede the old one and
is designed to hide all implementation details of EDL from the modules
(unlike the current version, which exposes JSON).