Move TOML Qt conversion to its own file and add more types
We can now serialize most of the basic Qt types to TOML. This should be
enough for most module's needs.
We do not support using integers, floats or any non-string type as keys
for tables, but that should not be a big limitation for modules (it's a
pretty obscure need anyway).
This paves the way for modules to use the markup language abstraction in
future, instead of hardcoding JSON as markup format.