Rename maio Python binding to syio
It's no longer associated with MazeAmaze and the API is so different now
that sharing the same module name with its ancestor doesn't make sense
any longer.
This also will trigger people to take a close look at their Python
scripts in Syntalos and adjust them to the new ports system.