firmata: Store millisecond timestamps in data and harmonize data properties
In the long run, we maybe want to give people more precise timing
control in general, e.g. by permitting to instruct a Firmata pin to be
on a certain value for an exact amount of time.
But for now, we at least store timestamps properly and convert them to
their Python representation with ease.