Change synchronizer logic to be insensitive to DAQ speed changes
A lot of cameras do not actually record with their designated framerate,
but with a slightly lower speed, e.g. 29.9Hz instead of 30Hz. Even these
small deviations threw off the previous algorithm. The new one shouldn't
care about the DAQ speed at all anymore because it doesn't predict
future time values but instead tries to ensure that the offset deviation
of the two clocks stays within bounds.

There's still lots of testing that needs to be done and some tweaks have
to be made (and this method may need to be used for the Intan
synchronizer as well).
All of that because some devices don't have support for hardware
synchronization...