Automatically increase thread priorities when running an experiment
This yields great latency improvements in some cases, while not hurting
the other usecases much.
And in general, increasing the priority of Syntalos threads over other
threads in the system seems like a good idea, since the computer will
likely be dedicated to running Syntalos anyway.
In case this feature causes issues or unwanted behavior, users are given
the option to disable it completely via global configuration changes.