camera-flir: Init & Deinit cam in same thread as Spn-System singleton
For some reason something in the Spinnaker code sends SIGKILL to its
thread on multiple initializations of the camera (SIGABRT would have
been much nicer to debug...).
The Spinnaker API documentation doesn't mention that any of the camera
functions are threadsafe or even reentrant, yet some of the Syntalos
code assumed reentrancy.
This change tries to mitigate the odd SIGKILL behavior by never
initializing, starting, stopping or destroying a camera in a different
thread than where the Spinnaker System singleton lives (and also where
the camera or the singleton where originally created).
There's also no hint in the docs that the Spinnaker System singleton is
actually threadsafe, or even shared between threads, so we couldn't just
assume that here.
Access to cam properties of a running camera from a different thread are
still permitted in this scheme, and so is (obviously) frame acquisition.
This seems to resolve the semi-random SIGKILL issue.