Try to give video encoding tasks a reasonable amount of threads to use
If we have spare threads that we are definitely not using, the video
encoder may have them. If we are using all of them, we can't really know
how busy the threads actually are, so in that case we give a default of
two threads to the encoder.
In any case, this should probably be adjustable by the user to fit their
needs.