Add experimenter selector for Syntalos boards
Blaming others for bad data got even easier, now that information about
who is responsible is stored directly with the experiment metadata.

This is useful for experiments which are conduced by multiple people or
across multiple groups, and simplifies finding out who created a dataset
without looking up that information in a lab notebook.