Make module driver selection a dedicated property and enum
It doesn't make sense to have these as flags, as you can never select
multiple modes at the same time (and you shouldn't be able to do that by
accident). Also, the module execution mode is kind of special, so it
makes sense to have that as its own property.