Simplify names of stored dataset by default
Store dataset names in their simple form. This is a very easy way to
avoid ambiguities and also issues due to spelling mistakes in filenames
or unescaped whitespaces in shell scripts.
Since this changes the storage layout, using simplified names is a
per-project setting and can be modified by the user.