Update OpalKelly FrontPanel BLOB
The newer versions don't load dynamically anymore, which is tricky for
us as the plan was to move the proprietary parts out of this
repository...
For now we at least update the shared library and see if it still works
with the old dynamically-loading C headers.
In the long run we need a better solution.