Reduce moc noise a bit
This is just band-aid. To fix this properly, we need to either wait for
Meson's automoc support to materialize, or manually separate out which
files need the moc to be run on them and which don't.
For the easy cases, this has been done, and for the bigger targets we
can get to it, so we don't run moc on files which don't need it and
generate hints and null-byte dummy files.
Getting rid of unnecessary moc'ing should also improve build speed.