Allow module UI drawing callbacks to be run directly
Using a QTimer adds a significant amount of delay and overhead, which
some modules which draw a lot of stuff can not afford (otherwise their
drawing may stutter, which doesn't make for a great UX).
Therefore, we add an option for these modules to get their redraw
functions called directly from the main loop during an experiment run.