intan-rhx: Work around main window not opening after save file reload
This only happens on real hardware without mock signal generation, and
also only happens when loading a saved board and not adding a "fresh"
module.
This has to be fixed properly, but in the meantime we just explicitly
hide modules that should be hidden when loading settings, which
mysteriously resolves the issue.