Don't allow user to interrupt preflight with a stop request
Syntalos can not stop the initial starting phase ("preflight") - if the
user issues a stop request during that phase, it confuses the state
machine and we end up with the application in a "disabled" state.
Users are impatient, so let's only allow them to stop a run if they
actually *can* stop it, and not before that (any lit up button will be
clicked...).
Preflight will either always succeed as a whole, or fail with an error.
Once all modules are prepared and the actual run has started, stopping
a run is possible again.