encodehelper: Prevent metadata race condition via global scheduled set
With the previous code, we only kept track of datasets submitted in
batch, which could still cause race conditions when multiple encoding
threads wanted to write the same metadata file at the same time.
We now keep a global record of directories we touched, and only clear it
if we finished processing all data.