Add feature to auto-repeat an experiment run at set intervals
This is mostly used for some QA and validation checks with various
configurations, to ensure Syntalos is keeping times correctly, but
besides for testing and validation this feature may be useful for other
purposes as well.
This commit exposes the feature in regular Syntalos builds in a somewhat
user-friendly way, with a GUI dialog.