Enable tuning flags for x86-64-v3, enable vectorization
This will tune for x86-64-v3 unconditionally, and enable features of
this level to optimize for it if desired.
For other architectures, we do nothing unless native optimizations are
enabled.
We also enable basic auto-vectorization now, which yields a small
benefit when transfering lots of video data and recording and processing
electrophysiology data.