timesync: Unconditionally write the last timestamp pair to tsync files
This change simplifies post-processing by removing the need for
excessive extrapolation when there is no sync event close to the end of
a recording. By always having a start and end time, we can now calculate
the time steps between the two, and remove a lot of guesswork.

Adding the extra data point also provides a nice tool to sanity-check
any previous data.