intan-rhx: Give save thread some time to write the last data to disk
If we do not wait a little, Syntalos will see signal data that then is
never written to disk, creating a mismatch between Syntalos' "live" view
on the generated data and the actual exported data.
This may for example result in timestamps being recorded by Syntalos
that exceed the amount of data stored to disk, which is not what we
want.