intan-rhx: Add hard barrier for tsync timestamps to not go out of bounds
This is obviously a hack, that prevents issues due to Syntalos seeing
timestamps that get never written into an Intan RHD file.
A better solution would require larger changes in the Intan RHX code,
which unfortunately is not desirable to keep rebasing on upstream's code
easy. In the long run, this can maybe be fixed properly upstream.