intan-rhx: Alter initial offset calculation to be more reliable
Since the master timestamp is meant to mean "the time when the last
sample was acquired", using FUNC_EXEC_TIMESTAMP to obtain it is not the
right way to do so. Using FUNC_DONE_TIMESTAMP is also not truly accurate
(due to the delay between the last sample acquisition and data
transmission that isn't accounted for in the device latency variable),
but using it is at least a bit better than before.

In order to get an accurate "initial time diff" between master clock and
secondary clock, we will also no longer use the error-prone first value
we get, but instead just export the expected offset value, which is a
lot more robust.