Clean up copyright headers, bump year
A lot of code was still marked as GPL-3.0 licensed even though we
switched to LGPL-3.0 as new default a while back, so while updating the
copyright year anyway, the per-file license also changed at least on
files which could be relicensed (there are a few belonging to multiple
copyright holders which remain GPL-3.0).
This means that the plugin interface code is now 100% LGPL-licensed, and
so are most of our plugins. The main application is still GPL-licensed.

This change allows plugins to pretty much choose whatever license they
like, without being bound by the GPL implicitly.
It also makes sharing code between the individual Syntalos components
easier.

This will also likely be the last yearly bump, doing so isn't of that
much value and makes Git history searches a bit more annoying.