canvas: Try even harder to resume image display when we are too slow
This is an odd issue which usually appeared when the graphics hardware
is faulty, but on one computer we actually now saw this on rapid
stop-and-restart cycles. For those events, we can maybe try a bit harder
to recover from a full "blackout", afterall Syntalos considers display
to be secondary and as long as the other modules are fine, the Canvas
can lag a bit behind.
If we lag around too much too frequently though, we need to stop as that
indicates a serious issue that the user should know about.