videorecorder: Copy frame before color conversion to prevent crash
Annoyingly, cv::Mat shares a backing memory pool, so a const cv::Mat
sets the surrounding structure as constant, but not the refcounted
actual data. So when we later perform work on the matrix, like color
conversions, there appears to be interference from other modules.
Copying the image fixes this issue, but it seems likely that we will
need to revisit this issue to make the buffer sharing more robust.
Potentially we can also make it easier to reason about who owns what by
using Eigen3 matrices instead of cv::Mat objects for images.