Add initial bits of Flatpak-awareness
In order for Syntalos to work well as a Flatpak, we need to add a bunch
of support code and checks for the host OS in order to make working with
this software pleasant from the sandboxed environment.
Knowing whether we're in a Flatpak sandbox or not is a good first step.