Make SysInfo a singleton
We are parsing quite a bit of information now when the structure is
created, and since all of it is unchanging over the program's lifetime,
making it a singleton is a useful optimization (especially since the
engine holds a reference to a SysInfo instance all the time already).
The SysInfo class is already threadsafe, but special care has to be
taken from now on to keep it that way.