Try to break thread deadlocks in misbehaving modules
This patch introduces a resilience measure by sending an empty message
to hanging threads that fail to join, in order to give potentially
waiting modules a chance to react to the termination request.

Since joining a std::thread from the outside using its handle confuses
the object (and crashes the app when it later unsuccessfully tries to
join) we created our own basic wrapper around pthread threads.

This feature is currently only used for module-specific threads, as the
issue should not happen for OOP workers and event-based modules.