Use VIPS as default image processing engine
While OpenCV is great overall, it is not very thread-save and cv::Mat is
not immutable by default. Furthermore, the pipeline approach of
VipsImage is extremely useful to create robust image processing
pipelines, and the overhead by cv::Mat vs VipsImage for video processing
was about the same when all data was in memory (data copying accounted
for most of the work anyway).

While the core is now passing around VipsImage entities for frames,
OpenCV is still there and converting between the two formats is very
straight-forward (unfortunately though we can not do this in a zero-copy
fashion, so some overhead will remain).