Initial refactoring step to make Syntalos' core a bit more modular
With the new IPC scheme we will soon have a lot more ways for modules to
be called, and we want to avoid any code duplication. This change is
just the first step to reorganize the source code to avoid dependency
conflicts and make the code easier to work with.