Rework streaming data type system
Using dynamic_cast or even qobject_cast to determine the data type at
hand is rather slow if it is done hundred-thousands of time in many
places at once.
In many places, we could skip type verification, but especially with the
new MLink system we will lose a lot of robustness.

This patch instead introduces a dedicated type system that provides a
unique typeId for every datatype while also adding new (de)serialization
methods that can be used efficiently with the Iceoryx IPC system that we
will switch the current home-grown IPC to.