Refactor and rework IPC interface and OOP module bindings
This is very early work to put our out-of-process modules onto a new
foundation. Instead of Qt Remote Objects and our own shmem IPC for when
QRO wasn't fast enough, we are now using Iceoryx, which, once fully
optimized, will improve the latency to OOP modules drmatically.

To accomodate the new changes, the Python API has also changed, and many
things are still not fully working, like the whole process of sending
settings to OOP modules. This will be added back soon, in a much better
way.
This is quite severe breakage, and existing projects will need to be
adjusted. But if we do this change now, and early, we will have a uch
more performant and flexible system in future, which could even permit
modules implemented in other, incompatible frameworks such as MATLAB or
Java.