Introduce a new module state for modules that don't participate in a run
Previously, we were overloading IDLE with double-meaning of both "the
module is at rest between two runs" and "the module is not participating
in the active run". This caused quite a few bugs and confusion, so let's
give modules a new state, DORMANT, to convey the information that they
purposefully decided to not participate in a run.