Make OpenCV matrix the default image-passing type again
While the memory management of OpenCV is a bit of an issue for Syntalos,
using VIPS turned out to require even more copies of image data than
thought. It also does create an annoying friction point due to OpenCV
being a lot more popular.
So, defaulting to OpenCV is the sensible, low-friction choice. We keep
VIPS around for now though, for operations where it excells, like
batch-processing multiple operations on an image.