camera-tis: Don't rely on the camera frame index and GST PTS
The GStreamer PTS isn't interesting for us, using the `capture_time_ns`
metadata is a lot more reliable.
In addition to that, the camera might drop the inditial frae, leading to
a frame count starting at 1 instead of 0, which messes up timestamping.
Instead, we create our own frame counter, counting only the frames that
we could actually use, instead of the ones handled by the camera.
That way, we get a counter with no gaps, as expected (if a frame was
dropped, that fact can still be inferred from the associated timestamp).