camera-tis: Don't average the initial clock offset again
It was a neat idea, but no big improvements where visible during tests
(but also no issues were found).
In addition to that though, there is a possibility that if the later
processing steps take too long and the internal buffers of the camera
pipeline are very full, we will get wrong time offsets.

Let's avoid that potential issue by going back to the previous, proven
code. We can add a smarter way for averaging back later, if needed and
proven beneficial.