videorecorder: Prevent deadlock if uninitialized module is stopped prematurely
This is a rare event, but we can't recover from it if it happens.
Usually this may occur if a device loses power early, failing the run,
and Syntalos sends a stop signal to a recorder module that never once
entered its prepare phase (because e.g. we just loaded a configuration
or the module was just added).

This patch resolves the problem.