camera-arv: udev: Disable powersaving on all USB cameras
This change also modernizes the rules a little bit. We still give it
both the uaccess tag and make the device world-read/writable, which is
quite silly. But I have no idea which stuff relies on the old rules, and
that's what the vendors also do, so let's keep it to make sure nothing
breaks.

All systems running Syntalos should have modern systemd, so we can drop
the udev-acl tag, at least.