canvas: Improve OpenGL shader compatibility
As dumb as it sounds, using GL 4.1 / GLES 3.2 seems to cause less
problems for people, likely because it is the better-tested codepath
now. We can likely bump this anyway, as any environment with Qt6 will
have a high enough GL version.
Also, just in case we hit a pedantic shader compiler, we now define the
precision values in regular GL as well (so far, every driver was fine
with the "wrong" shader, but you never know...).