camera-arv: Improve behavior if camera rescan is triggered by USB event
There were some really odd crashes in the QComboBox itself when a rescan
was triggered by a USB event. There was a way to simplify the code and
to reduce GUI churn if the background USB event has triggered a rescan,
so this patch implements the change. Which will hopefully also remove
the segmentation fault.

Sadly, the issue is not easily reproducible and may be in Qt itself, so
let's hope this works as workaround. In any case, the new code is much
cleaner and also re-selects the old camera after a rescan, which is much
nicer for the user.